continuing basic renderer implementation. swapchain and image views for drawing and presenting done

+++ include/Common.h
@@ -17,7 +17,7 @@
 using namespace std;
 using namespace glm;
 
-#define SCREEN_W 800
-#define SCREEN_H 600
+#define SCREEN_W (uint32_t)800
+#define SCREEN_H (uint32_t)600
 
 #endif // !_COMMON_H
